Created basic stuff for kingscup app
the name of the image doesn't fit to the rule of Android  so I created
another folder to put those "wrong" named images for poker.
Those should be named in the right way and put into the drawable
folder..

I will find some time to correct those 52 images file name.