Updated the Card and MainActivitiy classes and created a database helper class.
Card - Added a few instance variables to match up with the database I think we mapped out which is setup in the KingsCupDbAdapter class.

MainActivity - Added a changeImage method that will eventually be called to change the image to the correct image based off of the provided identifier.

KingsCupDbAdapter - A class that makes it easier to interact with the database from other classes. Add methods to it to make different queries as needed when programing.

As far as the database I came up with the idea of storing the R.drawable indentifiers in the database so all we would have to do if query with the cardname then switch the image to the correct one. So right now the database I have in my head is:

Database Name: kingscup
Table Name: cards
  Columns: _id (used as an guarented identifier. Required by several android classes)
           rdrawable (the R.drawable identifier for switching the images)
           cardname (basically the output of the toString method of the Card class)
           rule (the rule to display for the game)

hope that makes sense but if you have any edits you think we should make just let me know.